rename reporting batch resume router
