+++ src/app/router_registry.py
@@ -227,13 +227,11 @@ from app.routers.proposal_workflow_evidence import (
 )
 from app.routers.proposals import router as proposals_router
 from app.routers.reporting_batch_cancel import cancel_router as reporting_batch_cancel_router
-from app.routers.reporting_batch_controls import (
-    controls_router as reporting_batch_controls_router,
-)
 from app.routers.reporting_batch_lease_recovery import (
     recovery_router as reporting_batch_lease_recovery_router,
 )
 from app.routers.reporting_batch_pause import pause_router as reporting_batch_pause_router
+from app.routers.reporting_batch_resume import controls_router as reporting_batch_resume_router
 from app.routers.reporting_batch_retry import retry_router as reporting_batch_retry_router
 from app.routers.reporting_batch_run_once import (
     worker_router as reporting_batch_run_once_router,
@@ -439,7 +437,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(reporting_job_controls_router)
     app.include_router(reporting_batches_router)
     app.include_router(reporting_batch_status_router)
-    app.include_router(reporting_batch_controls_router)
+    app.include_router(reporting_batch_resume_router)
     app.include_router(reporting_batch_cancel_router)
     app.include_router(reporting_batch_pause_router)
     app.include_router(reporting_batch_retry_router)
